+++ core/src/test/scala/it/parttimeteam/core/GameManagerSuite.scala
@@ -6,10 +6,7 @@ import org.scalamock.matchers.Matchers
 import org.scalamock.scalatest.MockFactory
 import org.scalatest.funspec.AnyFunSpec
 import org.scalatest.matchers.must.Matchers.be
-import org.scalatest.matchers.should
 import org.scalatest.matchers.should.Matchers.an
-import org.scalatest.prop.{TableDrivenPropertyChecks, TableFor1}
-import org.scalatest.propspec.AnyPropSpec
 
 class GameManagerSuite extends AnyFunSpec with MockFactory with Matchers {
   describe("A game manager") {
@@ -139,38 +136,6 @@ class GameManagerSuite extends AnyFunSpec with MockFactory with Matchers {
         assert(!(hand.playerCards contains c1))
         assert(board.combinations forall (c => !(c.cards contains c1)))
       }
-    }
-  }
-}
-
-class GameManagerPropSpec extends AnyPropSpec with TableDrivenPropertyChecks with should.Matchers {
-  val gameManager: GameManager = new GameManagerImpl()
-
-  val c1: Card = Card.string2card("2♣R")
-  val c2: Card = Card.string2card("3♣B")
-  val c3: Card = Card.string2card("4♣B")
-  val c4: Card = Card.string2card("5♣B")
-  val c5: Card = Card.string2card("5♠R")
-  val c6: Card = Card.string2card("5♦B")
-  val c7: Card = Card.string2card("5♥R")
-
-  val validCombs: TableFor1[CardCombination] = Table("cardCombination",
-    CardCombination("#1", Seq(c1, c2, c3)),
-    CardCombination("#2", Seq(c4, c5, c6)),
-    CardCombination("#3", Seq(c4, c5, c6, c7)))
-
-  property("Test valid combinations") {
-    forAll(validCombs) { comb =>
-      gameManager validateCombination comb should be(true)
-    }
-  }
-
-  val invalidCombs: TableFor1[CardCombination] = Table("cardCombination",
-  CardCombination("#4", Seq(c4, c2, c7)))
-
-  property("Test invalid combinations") {
-    forAll(invalidCombs) { comb =>
-      gameManager validateCombination comb should be(false)
     }
   }
 }
