+++ core/src/test/scala/it/parttimeteam/core/cards/CardSuite.scala
@@ -50,50 +50,3 @@ class CardSuite extends AnyFunSuite {
     }
   }
 }
-
-object CardPropertySpec
-  extends AnyPropSpec
-    with TableDrivenPropertyChecks
-    with should.Matchers {
-
-  val ranks: TableFor1[Rank] = Table(
-    "rank",
-    Ace(),
-    Two(),
-    Three(),
-    Four(),
-    Five(),
-    Six(),
-    Seven(),
-    Eight(),
-    Nine(),
-    Ten(),
-    Jack(),
-    Queen(),
-    King()
-  )
-
-  val suits: TableFor1[Suit] = Table(
-    "suit",
-    Spades(),
-    Diamonds(),
-    Hearts(),
-    Clubs()
-  )
-
-  val tester: Card = Card(Rank.Three(), Suit.Spades(), Color.Red())
-
-  property("Check card compareto to another card with same suit.") {
-    forAll(ranks) { rank =>
-      val card = Card(rank, Suit.Spades(), Color.Red())
-      card compareTo tester should be(card.rank compareTo rank)
-    }
-  }
-
-  property("Check card compareto to another card with different suit") {
-    forAll(suits) { suit =>
-      val card = Card(Rank.Five(), suit, Color.Blue())
-      card compareTo tester should be(card.suit compareTo suit)
-    }
-  }
-}
