+++ core/src/test/scala/it/parttimeteam/core/cards/SuitSuite.scala
@@ -2,9 +2,6 @@ package it.parttimeteam.core.cards
 
 import it.parttimeteam.core.cards.Suit._
 import org.scalatest.funsuite.AnyFunSuite
-import org.scalatest.matchers.should
-import org.scalatest.prop.{TableDrivenPropertyChecks, TableFor1}
-import org.scalatest.propspec.AnyPropSpec
 
 class SuitSuite extends AnyFunSuite {
   test("Check all suits properties") {
@@ -23,49 +20,5 @@ class SuitSuite extends AnyFunSuite {
 
   test("Check all suits") {
     assert(Suit.all equals List(Hearts(), Diamonds(), Clubs(), Spades()))
-  }
-}
-
-class SuitPropSpec
-  extends AnyPropSpec
-    with TableDrivenPropertyChecks
-    with should.Matchers {
-
-  // Property test for all suits.
-  val examples: TableFor1[Suit] = Table(
-    "suit",
-    Clubs(),
-    Spades(),
-    Diamonds(),
-    Hearts()
-  )
-
-  val failures: TableFor1[String] = Table(
-    "suit",
-    "s",
-    "A",
-    "J",
-    "Q",
-    "K"
-  )
-
-  property("string2suit must return shortName at name input") {
-    forAll(examples) { suit =>
-      (Suit string2suit suit.shortName).name should be(suit.name)
-    }
-  }
-
-  property("Invoking string2suit with other strings must return RuntimeException") {
-    forAll(failures) { fail =>
-      a[RuntimeException] should be thrownBy {
-        Suit string2suit fail
-      }
-    }
-  }
-
-  property("Check comparation between suits") {
-    forAll(examples) { suit =>
-      suit compareTo Diamonds() should be(suit.order compareTo Diamonds().order)
-    }
   }
 }
